requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
google-auth>=2.14.1,!=2.24.0,!=2.25.0,<3.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2>=0.2.0,<1.0.0
requests==2.31.0
redis==5.0.1